requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backoff==1.10.0
bcrypt==3.2.0
bpemb==0.3.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
conllu==4.4.1
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.23
Deprecated==1.2.13
doc2text==0.2.4
fastapi==0.73.0
ffmpy==0.3.0
filelock==3.4.2
flair==0.10
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.3.0
ftfy==6.0.3
future==0.18.2
gdown==3.12.2
gensim==4.1.2
gradio==2.7.5.2
h11==0.13.0
huggingface-hub==0.4.0
humanfriendly==10.0
idna==3.3
importlib-metadata==3.10.1
Janome==0.4.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
markdown2==2.4.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mime==0.1.0
monotonic==1.6
more-itertools==8.8.0
mpld3==0.3
multidict==6.0.2
numpy==1.23.3
overrides==3.1.0
packaging==21.3
pandas==1.4.0
paramiko==2.9.2
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.1
protobuf==3.19.4
psutil==5.9.0
py-cpuinfo==8.0.0
py3nvml==0.2.7
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn
scipy==1.7.3
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sqlitedict==1.7.0
starlette==0.17.1
tabulate==0.8.9
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.10.2
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
Wand==0.6.7
wcwidth==0.2.5
Wikipedia-API==0.5.4
wincertstore==0.2
wrapt==1.13.3
xmltodict==0.12.0
yarl==1.7.2
zipp==3.7.0
openai
sqlite3
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib